pyproject.toml
[tool.poetry]
name = "likelyindolentcharacter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^1.1.2"
praw = "^7.1.0"
python-dotenv = "^0.15.0"
discord = "^1.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
